feat(artifacts): Re-use artifact configuration
