+++ config/artifact_repository.go
@@ -1,0 +1,69 @@
+package config
+
+import wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
+
+// ArtifactRepository represents a artifact repository in which a controller will store its artifacts
+type ArtifactRepository struct {
+	// ArchiveLogs enables log archiving
+	ArchiveLogs *bool `json:"archiveLogs,omitempty"`
+	// S3 stores artifact in a S3-compliant object store
+	S3 *S3ArtifactRepository `json:"s3,omitempty"`
+	// Artifactory stores artifacts to JFrog Artifactory
+	Artifactory *ArtifactoryArtifactRepository `json:"artifactory,omitempty"`
+	// HDFS stores artifacts in HDFS
+	HDFS *HDFSArtifactRepository `json:"hdfs,omitempty"`
+	// OSS stores artifact in a OSS-compliant object store
+	OSS *OSSArtifactRepository `json:"oss,omitempty"`
+	// GCS stores artifact in a GCS object store
+	GCS *GCSArtifactRepository `json:"gcs,omitempty"`
+}
+
+func (a *ArtifactRepository) IsArchiveLogs() bool {
+	return a != nil && a.ArchiveLogs != nil && *a.ArchiveLogs
+}
+
+// S3ArtifactRepository defines the controller configuration for an S3 artifact repository
+type S3ArtifactRepository struct {
+	wfv1.S3Bucket `json:",inline"`
+
+	// KeyFormat is defines the format of how to store keys. Can reference workflow variables
+	KeyFormat string `json:"keyFormat,omitempty"`
+
+	// KeyPrefix is prefix used as part of the bucket key in which the controller will store artifacts.
+	// DEPRECATED. Use KeyFormat instead
+	KeyPrefix string `json:"keyPrefix,omitempty"`
+}
+
+// OSSArtifactRepository defines the controller configuration for an OSS artifact repository
+type OSSArtifactRepository struct {
+	wfv1.OSSBucket `json:",inline"`
+
+	// KeyFormat is defines the format of how to store keys. Can reference workflow variables
+	KeyFormat string `json:"keyFormat,omitempty"`
+}
+
+// GCSArtifactRepository defines the controller configuration for a GCS artifact repository
+type GCSArtifactRepository struct {
+	wfv1.GCSBucket `json:",inline"`
+
+	// KeyFormat is defines the format of how to store keys. Can reference workflow variables
+	KeyFormat string `json:"keyFormat,omitempty"`
+}
+
+// ArtifactoryArtifactRepository defines the controller configuration for an artifactory artifact repository
+type ArtifactoryArtifactRepository struct {
+	wfv1.ArtifactoryAuth `json:",inline"`
+	// RepoURL is the url for artifactory repo.
+	RepoURL string `json:"repoURL,omitempty"`
+}
+
+// HDFSArtifactRepository defines the controller configuration for an HDFS artifact repository
+type HDFSArtifactRepository struct {
+	wfv1.HDFSConfig `json:",inline"`
+
+	// PathFormat is defines the format of path to store a file. Can reference workflow variables
+	PathFormat string `json:"pathFormat,omitempty"`
+
+	// Force copies a file forcibly even if it exists (default: false)
+	Force bool `json:"force,omitempty"`
+}
